faceguard-v2/services/face-recognition-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
numpy==1.24.3
faiss-cpu==1.7.4
psutil==5.9.6
python-multipart==0.0.6